back/requirements.txt
django==3.2.20
djangorestframework
pillow
setuptools
django-cors-headers
drf-yasg